requirements.txt
nodejs
yt_dlp
flask
flask-cors
requests
urllib3
pymongo
discord.py 
motor 
python-dotenv
aiohttp